feat: implement exception handling and vector table setup
- Added exception vector table in `exception.s` and linked it in `linker.ld`
- Implemented `set_exception_vector_table` to set `vbar_el1`
- Implemented exception handlers:
  - `_el1_lower_el_aarch64_sync_handler`: logs `ELR_EL1`, `SPSR_EL1`, and `ESR_EL1`
  - `_el1_current_el_aarch64_sync_handler` (stub)
  - `_el1_current_el_aarch64_irq_handler` (stub)
  - `_el1_lower_el_aarch64_irq_handler` (stub)
  - `default_exception_handler` for unhandled exceptions
- Updated `boot.s` to set up the exception vector table during boot

+++ kernel/core/command.c
@@ -1,12 +1,12 @@
-#include "fs/file.h"
+#include "common/types.h"
+#include "common/utils.h"
+#include "core/simple_alloc.h"
+#include "core/user_exec.h"
 #include "drivers/mailbox.h"
 #include "drivers/mbox_tags.h"
-#include "core/simple_alloc.h"
-#include "common/types.h"
 #include "drivers/uart.h"
-#include "core/user_exec.h"
-#include "common/utils.h"
 #include "drivers/watchdog.h"
+#include "fs/file.h"
 
 void cmd_help(unsigned int argc, const char *argv[]) {
   uart_send_string("help      : print this help menu\r\n"
@@ -138,7 +138,7 @@ void cmd_mem_alloc(unsigned int argc, const char *argv[]) {
   }
   uart_send_string("Allocated memory at: 0x");
   char hexstr_buf[9] = {};
-  uint2hex((unsigned int)ptr, hexstr_buf);
+  uint2hex((uint64_t)ptr, hexstr_buf);
   uart_send_string(hexstr_buf);
   uart_send_string("\r\n");
 }
